shortend mediaPost code
